+++ packages/context/__tests__/shared.js
@@ -40,8 +40,13 @@ export default () => {
     const testInstance = TestRenderer.create(
       <SectionContext.Provider
         value={{
+          onArticleSavePress: () => {},
           publicationName: "SUNDAYTIMES",
-          recentlyOpenedPuzzleCount: 123
+          recentlyOpenedPuzzleCount: 123,
+          savedArticles: {
+            "dummy-article-id-1": true,
+            "dummy-article-id-2": true
+          }
         }}
       >
         <SectionContext.Consumer>