+++ packages/edition-slices/src/slices/standard/styles.js
@@ -1,0 +1,11 @@
+import { spacing } from "@times-components/styleguide";
+
+const mediumBreakpointStyles = {
+  container: {
+    flex: 1,
+    flexDirection: "row",
+    paddingHorizontal: spacing(10)
+  }
+};
+
+export default mediumBreakpointStyles;