+++ packages/pages/__tests__/mocks.js
@@ -27,7 +27,9 @@ jest.mock("react-native", () => {
   };
   rn.NativeModules.SectionEvents = {
     getOpenedPuzzleCount: jest.fn(),
+    getSavedArticles: jest.fn().mockReturnValue(Promise.resolve([])),
     onArticlePress: () => {},
+    onArticleSavePress: jest.fn().mockReturnValue(Promise.resolve(true)),
     onPuzzleBarPress: () => {},
     onPuzzlePress: () => {},
     onSectionLoaded: () => {}