+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.75](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.74...@times-components/topic@4.3.75) (2019-04-02)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [4.3.74](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.73...@times-components/topic@4.3.74) (2019-04-01)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [4.3.73](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.72...@times-components/topic@4.3.73) (2019-04-01)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [4.3.72](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.71...@times-components/topic@4.3.72) (2019-04-01)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [4.3.71](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.70...@times-components/topic@4.3.71) (2019-04-01)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 ## [4.3.70](https://github.com/newsuk/times-components/compare/@times-components/topic@4.3.69...@times-components/topic@4.3.70) (2019-04-01)
 
 **Note:** Version bump only for package @times-components/topic